// src/types.ts
export interface CanvasElements {
	boardCanvas: HTMLCanvasElement;
	holdCanvas: HTMLCanvasElement;
	queueCanvas: HTMLCanvasElement;
	bestMoveCanvas: HTMLCanvasElement;
	boardCtx: CanvasRenderingContext2D;
	holdCtx: CanvasRenderingContext2D;
	queueCtx: CanvasRenderingContext2D;
	bestMoveCtx: CanvasRenderingContext2D;
}